+++ package/console/create/dkml_package_console_create.mli
@@ -1,4 +1,5 @@
 val create_installers :
+  Dkml_package_console_common.Author_types.build_info ->
   Dkml_package_console_common.Author_types.organization ->
   Dkml_package_console_common.Author_types.program_name ->
   Dkml_package_console_common.Author_types.program_info ->