+++ runner/bin/dkml_install_runner_admin.mli
@@ -1,2 +1,5 @@
 val main :
-  target_abi:Dkml_install_api.Context.Abi_v2.t -> program_version:string -> unit
+  target_abi:Dkml_install_api.Context.Abi_v2.t ->
+  package_name:string ->
+  program_version:string ->
+  unit